// src/shared/api/models/postApiBundleOffersBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiBundleOffersBodyApplicableService } from './postApiBundleOffersBodyApplicableService';
import type { PostApiBundleOffersBodyBundleOfferType } from './postApiBundleOffersBodyBundleOfferType';

export type PostApiBundleOffersBody = {
    /**
     * Name of the bundle offer
     * @minLength 1
     * @maxLength 255
     */
    name: string;
    /** Description of the bundle offer */
    description?: string;
    /** Services on which bundle offer can be applied */
    applicableService?: PostApiBundleOffersBodyApplicableService;
    /** Type of bundle offer */
    bundleOfferType?: PostApiBundleOffersBodyBundleOfferType;
    /**
     * Price of the bundle offer
     * @minimum 0
     */
    price?: number;
    /**
     * Number of months until expiry
     * @minimum 1
     */
    expiryMonths?: number;
    /** Whether to sell online */
    sellOnline?: boolean;
};